Refactor Auditor to simplify the readability
